chore: move types out of config.rs into config_types.rs
