+++ codex-rs/core/src/lib.rs
@@ -13,6 +13,7 @@ pub use codex::Codex;
 pub mod codex_wrapper;
 pub mod config;
 pub mod config_profile;
+pub mod config_types;
 mod conversation_history;
 pub mod error;
 pub mod exec;
@@ -22,7 +23,6 @@ mod is_safe_command;
 #[cfg(target_os = "linux")]
 pub mod landlock;
 mod mcp_connection_manager;
-pub mod mcp_server_config;
 mod mcp_tool_call;
 mod message_history;
 mod model_provider_info;
